requirements.txt
PyGithub>=2.1.0
openai>=1.12.0
Jinja2>=3.1.3
requests>=2.31.0
